feat: remove the extraneous random data from the JWT payload
That data was added to make sure the tokens where unique when renewing.
In fact the only situation where that could happen was in the test where
the token was issued and renewed almost simultaneously. To avoid that
a delay has been added to the TEST between issuing the original token and
renewing it.